Replace the temporary fix for the ingredients; "Water" and "Beans"
The ingredients "Water" and "Beans", required additional properties, "Temperature" and "CoffeSorts" respectively.
This meant that the way we defined the ingredients "Water" and "Beans" required an update.
This change was implemented in commit 410b0a8.

We decided to define "Water" and "Beans" in separate classes, and a temporary fix we added these "new" ingredients to the EspressoMachine-class as separate properties.
Another temporary fix was to add water and beans to the "ingredientsSorted"-collection manually in "EspressoMachine.ToBeverage()".

This commit will replace these temporary fixes.
By having the classes "Water" and "Beans" inherit from the "Ingredient"-class, we can add "Water" and "Beans" objects to the "Ingredients"-List in the EspressoMachine-class.
And the temporary fixes introduced in commit 410b0a8 can be removed/reverted.