Set a default for the Name property on the Water and Beans classes
By using "virtual" on the "Name"-property in the" Ingredients"-class we can override the property on any class that inherits from the "Ingredients"-class.
We can thus remove the code settings this property where we create new "Water" and "Beans" objects (in the "AddWater()"-method and in the API calls in "Program.cs").